refactor: separate makeShader
