+++ src/gfx/shader.ts
@@ -1,0 +1,36 @@
+import {
+    DEF_FRAG,
+    DEF_VERT,
+    FRAG_TEMPLATE,
+    VERT_TEMPLATE,
+    VERTEX_FORMAT,
+} from "../constants";
+import { type GfxCtx, Shader } from "../gfx";
+import { getErrorMessage } from "../utils";
+
+export function makeShader(
+    ggl: GfxCtx,
+    vertSrc: string | null = DEF_VERT,
+    fragSrc: string | null = DEF_FRAG,
+): Shader {
+    const vcode = VERT_TEMPLATE.replace("{{user}}", vertSrc ?? DEF_VERT);
+    const fcode = FRAG_TEMPLATE.replace("{{user}}", fragSrc ?? DEF_FRAG);
+
+    try {
+        return new Shader(
+            ggl,
+            vcode,
+            fcode,
+            VERTEX_FORMAT.map((vert) => vert.name),
+        );
+    } catch (e) {
+        const lineOffset = 14;
+        const fmt = /(?<type>^\w+) SHADER ERROR: 0:(?<line>\d+): (?<msg>.+)/;
+        const match = getErrorMessage(e).match(fmt);
+        if (!match?.groups) throw e;
+        const line = Number(match.groups.line) - lineOffset;
+        const msg = match.groups.msg.trim();
+        const ty = match.groups.type.toLowerCase();
+        throw new Error(`${ty} shader line ${line}: ${msg}`);
+    }
+}
